feat: Implemented resource owner check - temporal, added e2e tests, added auto-populate to ClaimSchema

+++ src/claims/claims.service.ts
@@ -1,5 +1,10 @@
 import { _ } from 'lodash';
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Claim, ClaimDocument } from './schemas/claim.schema';
 import { InjectModel } from '@nestjs/mongoose';
 import { User } from '../users/schemas/user.schema';
@@ -10,6 +15,28 @@ import { NullableType } from 'src/utils/types/nullable.type';
 @Injectable()
 export class ClaimsService {
   constructor(@InjectModel(Claim.name) private claimModel: Model<Claim>) {}
+
+  async checkResourceAccess(user: User, _id: Types.ObjectId): Promise<boolean> {
+    if (_.includes(user.roles, 'admin')) return true;
+
+    const claim: Claim = await this.claimModel.findOne({ _id });
+
+    if (!claim) {
+      throw new NotFoundException('Claim not found');
+    }
+
+    if (!_.isEqual(claim.addedBy._id, user._id)) {
+      throw new HttpException(
+        {
+          statusCode: HttpStatus.FORBIDDEN,
+          message: 'Forbidden',
+        },
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
+    return true;
+  }
 
   async create(
     articleId: Types.ObjectId,
@@ -23,10 +50,12 @@ export class ClaimsService {
         articles: [articleId],
       }),
     );
+    // TODO add claimId to Article{articleId}.claims
     return createdClaim.save();
   }
 
   async findOne(query: object): Promise<NullableType<Claim>> {
+    // TODO add review by current user if exists
     const claim = await this.claimModel.findOne(query);
     if (!claim) {
       throw new NotFoundException(`Claim not found`);
@@ -39,6 +68,7 @@ export class ClaimsService {
     page = 1,
     perPage?: number,
   ): Promise<Claim[]> {
+    // TODO add current user's reviews
     return this.claimModel
       .find({ article: articleId })
       .skip(perPage * (page - 1))
@@ -49,6 +79,7 @@ export class ClaimsService {
   //   _id: Types.ObjectId,
   //   updateClaimDto: UpdateClaimDto,
   // ): Promise<Claim> {
+  // await this.checkResourceAccess(loggedUser, claimId);
   //   const updatedClaim: Claim = await this.claimModel.findByIdAndUpdate(
   //     _id,
   //     updateClaimDto,
@@ -67,7 +98,8 @@ export class ClaimsService {
   //   loggedUser: User,
   //   claimDto: ReplaceClaimDto,
   // ): Promise<Claim> {
-  //   return await this.claimModel.findByIdAndUpdate(
+  // await this.checkResourceAccess(loggedUser, claimId);
+  //   return this.claimModel.findByIdAndUpdate(
   //     _id,
   //     _.assign(claimDto, { addedBy: loggedUser._id }),
   //     {
@@ -79,9 +111,12 @@ export class ClaimsService {
   // }
 
   async delete(
+    loggedUser: User,
     articleId: Types.ObjectId,
     claimId: Types.ObjectId,
   ): Promise<Claim> {
+    await this.checkResourceAccess(loggedUser, claimId);
+
     const deleterClaim = await this.claimModel.findByIdAndDelete({
       article: articleId,
       _id: claimId,
